requirements.txt
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
cloudinary==1.44.0
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-flask==1.3.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
WTForms==3.2.1
